services/llm_parser_service/requirements.txt
minio>=7.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
gigachat
pydantic-settings
langchain-gigachat